# Make `oj_dump_ignore()` an inline function