To improve `Oj.dump` performance, this PR make `oj_dump_ignore()` an inline function.

−               | before | after  | result
--               | --     | --     | --
Oj.dump          | 1.190M | 1.216M | 1.022x

### Environment
- MacBook Pro (M1 Max, 2021)
- macOS 12.0
- Apple M1 Max
- Ruby 3.0.2

### Before
```
Warming up --------------------------------------
             Oj.dump   119.103k i/100ms
Calculating -------------------------------------
             Oj.dump      1.190M (± 0.7%) i/s -     11.910M in  10.009672s
```

### After
```
Warming up --------------------------------------
             Oj.dump   121.394k i/100ms
Calculating -------------------------------------
             Oj.dump      1.216M (± 0.5%) i/s -     12.261M in  10.086493s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'

json =<<-EOF
{
  "$id": "https://example.com/person.schema.json",
  "$schema": "https://json-schema.org/draft/2020-12/schema",
  "title": "Person",
  "type": "object",
  "properties": {
    "firstName": {
      "type": "string",
      "description": "The person's first name."
    },
    "lastName": {
      "type": "string",
      "description": "The person's last name."
    },
    "age": {
      "description": "Age in years which must be equal to or greater than zero.",
      "type": "integer",
      "minimum": 0
    }
  }
}
EOF

Benchmark.ips do |x|
  x.warmup = 10
  x.time = 10

  data = Oj.load(json, symbol_keys: true)
  x.report('Oj.dump') { Oj.dump(data) }
end
```